added capability of pruning elite set on decision variables
Previously, pruning was done on objective function values.  In the
case where that is not possible (e.g. cannot find difference of
values), we consider the decision variables as well.  The latter
assume that we have a -(subtraction) operator for the decision
variable type; if not, we do no pruning at all.